requirements.txt
fastapi==0.109.0
greenlet==3.0.3
uvicorn==0.27.0
SQLAlchemy==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
python-dotenv==1.0.1
minio==7.2.7